Fixed Safari Display for Logo
This page uses Bootstrap.

This version has a home image as a background with text.

This scales well on a mobile device. It has a menu that is hidden, but can be displayed with a menu button.

Google Font Montserrat has been added, bold is enabled, more format adjustments.

Custom colors and other design elements added.

Temporarily removed Footer to check mobile design.

Updated logo design to work in Safari as well as Chrome.